split up daemon into v1/v2 routes
What I Did
------------

split up daemon into v1/v2 routes

This "rework-things to support regular nav", is turning out to be a huge project,
so I'm going to copy stuff over from that other branch by hand, but it has a ton of
breaking changes that I want to try to avoid.

How I Did it
------------

add `V1Routes` and `V2Routes` struct, daemon inherits the routes from these. Someday maybe we'll be able to
ditch V1 altogether, but I want to try and keep these PRs small and this seems a good way to do it.

How to verify it
------------

Should be no visible changes here, except the `lifecycle` route at `/api/v2/lifecycle`

:rowboat:
