+++ pkg/lifecycle/daemon/daemon_channel_test.go
@@ -64,13 +64,17 @@ func TestDaemonChannel(t *testing.T) {
 			fs := afero.Afero{Fs: afero.NewMemMapFs()}
 			log := &logger.TestLogger{T: t}
 			daemon := &ShipDaemon{
-				Logger: log,
-				Fs:     fs,
-				Viper:  v,
+				Logger:       log,
+				WebUIFactory: WebUIFactoryFactory(log),
+				Viper:        v,
+				V1Routes: &V1Routes{
+					Logger: log,
+					Fs:     fs,
+					Viper:  v,
 
-				UI:             cli.NewMockUi(),
-				WebUIFactory:   WebUIFactoryFactory(log),
-				OpenWebConsole: func(ui cli.Ui, s string) error { return nil },
+					UI:             cli.NewMockUi(),
+					OpenWebConsole: func(ui cli.Ui, s string) error { return nil },
+				},
 			}
 
 			daemonCtx, daemonCancelFunc := context.WithCancel(context.Background())
